server/pyproject.toml
[project]
name = "phone-ag-server"
version = "0.1.0"
description = "AI phone agent server using LiveKit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit>=0.13.0",
    "livekit-agents>=0.10.0",
    "livekit-plugins-openai>=0.9.0",
    "livekit-plugins-deepgram>=0.6.0",
    "livekit-plugins-cartesia>=0.3.0",
    "livekit-plugins-silero>=0.6.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "redis>=5.2.0",
    "twilio>=9.4.0",
    "python-multipart>=0.0.20",
    "websockets>=14.2",
    "psutil>=6.1.1",
    "aiofiles>=24.1.0",
    "firebase-admin>=6.5.0",
    "google-cloud-firestore>=2.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "black>=24.10.0",
    "ruff>=0.8.6",
    "mypy>=1.14.2",
]
